% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_pp.R
\name{glm_pp}
\alias{glm_pp}
\title{Power prior with fixed a0 for generalized linear models using Stan}
\usage{
glm_pp(
  formula,
  family,
  data,
  histdata,
  a0 = 0.5,
  beta0 = NULL,
  Sigma0 = NULL,
  offset = NULL,
  offset0 = NULL,
  disp.shape = 2.1,
  disp.scale = 1.1,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}.}

\item{family}{an object of type \code{\link[stats]{family}} giving distribution and link function}

\item{data}{a \code{\link[base]{data.frame}} of current data giving all variables in \code{formula}}

\item{histdata}{a \code{\link[base]{data.frame}} of historical data giving all variables in \code{formula}}

\item{a0}{fixed power prior parameter}

\item{beta0}{mean for initial prior on regression coefficients. Defaults to vector of 0s}

\item{Sigma0}{covariance matrix for initial prior on regression coefficients. Defaults to \code{diag(100, ncol(X))}}

\item{offset}{offset for current data in GLM. If \code{NULL}, no offset is utilized}

\item{offset0}{offset for historical data in GLM. If \code{NULL}, no offset is utilized}

\item{disp.shape}{shape parameter for inverse-gamma prior on dispersion (for Gaussian and gamma models). Ignored for binomial and Poisson models}

\item{disp.scale}{scale parameter for inverse-gamma prior on dispersion (for Gaussian and gamma models). Ignored for binomial and Poisson models}

\item{...}{optional parameters to pass onto \code{rstan::sampling}}
}
\value{
an object of class \link[rstan:stanfit-class]{rstan::stanfit} returned by \code{rstan::sampling}
}
\description{
Sample from the posterior of a power prior (PP) generalized linear model
(GLM) with a fixed a0.
}
\examples{
## Generate current and historical data
set.seed(123)
n  = 30
n0 = 20
x  = rnorm(n)
x0 = rnorm(n0)
y  = rbinom(n = n,  size = 1, prob = binomial()$linkinv(1 - 0.5 * x) )
y0 = rbinom(n = n0, size = 1, prob = binomial()$linkinv(1 - 0.5 * x0) )
data = data.frame('y' = y, 'x' = x)
histdata = data.frame('y' = y0, 'x' = x0)

## Sample from posterior distribution of non-normalized power prior
fit.pp = glm_pp(y ~ x, binomial(), data, histdata, a0 = 0.5)
}
