% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_npp.R
\name{lm_npp}
\alias{lm_npp}
\title{Normalized power prior for for the normal linear model (identity link) using Stan}
\usage{
lm_npp(
  formula,
  data,
  histdata,
  beta0 = NULL,
  Sigma0 = NULL,
  offset = NULL,
  offset0 = NULL,
  sigmasq.shape = 2.1,
  sigmasq.rate = 1.1,
  a0.shape1 = 1,
  a0.shape2 = 1,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}.}

\item{data}{a \code{\link[base]{data.frame}} of current data giving all variables in \code{formula}}

\item{histdata}{a \code{\link[base]{data.frame}} of historical data giving all variables in \code{formula}}

\item{beta0}{mean for initial prior on regression coefficients. Defaults to vector of 0s}

\item{Sigma0}{covariance matrix for initial prior on regression coefficients. Defaults to \code{diag(100, ncol(X))}}

\item{offset}{offset for current data. If \code{NULL}, no offset is utilized}

\item{offset0}{offset for historical data. If \code{NULL}, no offset is utilized}

\item{sigmasq.shape}{shape parameter for inverse-gamma prior on variance parameter. When \code{sigmasq.shape==2.1} and \code{sigmasq.rate==1.1}, a prior with mean 1 and variance 100 is utilized}

\item{sigmasq.rate}{rate parameter for inverse-gamma prior on variance parameter. When \code{sigmasq.shape==2.1} and \code{sigmasq.rate==1.1}, a prior with mean 1 and variance 100 is utilized}

\item{a0.shape1}{shape1 parameter for beta prior on the power prior parameter. When \code{a0.shape1==1} and \code{a0.shape2==1}, a uniform prior is utilized}

\item{a0.shape2}{shape2 parameter for beta prior on the power prior parameter. When \code{a0.shape1==1} and \code{a0.shape2==1}, a uniform prior is utilized}

\item{...}{optional parameters to pass onto \code{rstan::sampling}}
}
\value{
an object of class \link[rstan:stanfit-class]{rstan::stanfit} returned by \code{rstan::sampling}
}
\description{
Sample from the posterior of a normalized power prior (NPP) of a linear model (LM), where the normalizing constant is known.
}
\examples{
## Generate current and historical data
set.seed(123)
N = 100
N0 = 50
x = rnorm(N)
y = 1 + 0.5 * x + rnorm(N)
x0 = rnorm(N0)
y0 = 1 + 0.5 * x0 + rnorm(N0)
data = data.frame('y' = y, 'x' = x)
histdata = data.frame('y' = y0, 'x' = x0)

## Perform MCMC of normalized power prior for normal linear model
fit = lm_npp(y ~ x, data, histdata)

}
